Revert package.json and move jsonata tests into separate file
